+++ packages/implementations/react/src/atomic/Banner/src/index.ts
@@ -1,0 +1,5 @@
+import { Banner } from "./Banner";
+
+export { Banner } from "./Banner";
+export type { BannerProps } from "./banner.types";
+export default Banner;