+++ packages/styles/src/index.ts
@@ -14,6 +14,7 @@ export { skeleton } from "./packages/atomic/skeleton";
 export { status } from "./packages/atomic/status";
 export { text } from "./packages/atomic/text";
 export { title } from "./packages/atomic/title";
+export { banner } from "./packages/atomic/banner";
 export { toast } from "./packages/atomic/toast";
 export { toggle } from "./packages/atomic/toggle";
 
@@ -30,5 +31,6 @@ export type { CheckboxVariants } from "./packages/atomic/checkbox";
 export type { ChipVariants } from "./packages/atomic/chip";
 export type { IconButtonVariants } from "./packages/atomic/iconButton";
 export type { StatusVariants } from "./packages/atomic/status";
+export type { BannerVariants } from "./packages/atomic/banner";
 export type { ToastVariants } from "./packages/atomic/toast";
 export type { ToggleVariants } from "./packages/atomic/toggle";