+++ packages/styles/src/packages/atomic/banner/rarui-banner.types.ts
@@ -1,0 +1,4 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { styles } from "./raui-banner.css";
+
+export type BannerVariants = RecipeVariants<typeof styles.banner>;